Pulled up the variable InMemoGenericTagStorage storage to the class AbstarctSTSet.java + Pulled up the methods addGeoSemanticTag(), createSpatialSemanticTag(), castGST(), getSpatialSemanticTag() to the class AbstarctSTSet.java
